Bugfix di TabooConfigurat* (i due parametri per la packing procedure ora sono realmente float) C'è un IllegalArgumentException in giro, probabilmente dovuta al fatto che i pacchetti non vengono gestiti bene

+++ iaproject-2dbpp/src/core/taboo/TabooConfiguration.java
@@ -7,15 +7,15 @@ public class TabooConfiguration implements Serializable {
 	private static final long serialVersionUID = -1688023669983960748L;
 	
 	public final float ALPHA;
-	public final int DENSITY_FACTOR;
-	public final int HEIGHT_FACTOR;
+	public final float DENSITY_FACTOR;
+	public final float HEIGHT_FACTOR;
 	public final int MAX_NEIGH_SIZE;
 	public final int D_MAX;
 	public final int FIRST_LIST_TENURE;
 	public final int OTHER_LIST_TENURE;
 	
-	public TabooConfiguration(float aLPHA, int dENSITY_FACTOR,
-			int hEIGHT_FACTOR, int mAX_NEIGH_SIZE, int d_MAX,
+	public TabooConfiguration(float aLPHA, float dENSITY_FACTOR,
+			float hEIGHT_FACTOR, int mAX_NEIGH_SIZE, int d_MAX,
 			int fIRST_LIST_TENURE, int oTHER_LIST_TENURE) {
 		ALPHA = aLPHA;
 		DENSITY_FACTOR = dENSITY_FACTOR;
